6/5/2025
save game data

+++ Assets/Scripts/SaveGame/SaveData.cs
@@ -3,9 +3,9 @@ public class SaveData
 {
     public float playerX;
     public float playerY;
-    public int currentHealth;
-    public int maxHealth;
-    public int currentEnergy;
+    public float currentHealth;
+    public float maxHealth;
+    public float currentEnergy;
     public string currentMap;
     public string currentForm; // ví dụ: "default", "dark"
 }
